requirement.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/ML
langchain==0.1.0
openai==1.6.1
chromadb==0.4.18
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.36.2


# database
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1

# semantic
rdflib==7.0.0
owlready2==0.45
SPARQLWrapper==2.0.0

# tools
python-dotenv==1.0.0
pyyaml==6.0.1
pandas==2.1.4
numpy==1.26.2
structlog==23.2.0

# auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# test
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# monitor
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.1